# HTML API: Optimize low-level parsing details.