Trac ticket: Core-61545

## Status

 - passes all tests, code audit looks right
 - performs more efficiently in almost all cases, but worse in only a few minor cases. the expected speedup is around 3.5% - 7.5% for the cost of parsing the tags in realistic HTML documents.

## Summary

Measured change in parsing time to count tags in HTML5 spec:
 - from 1050ms to 930ms
 - roughly 13% faster in the HTML spec `single-page.html`
 - much less speedup in small documents with a higher text-to-syntax ratio

## Benchmarking

Based on the `top100` set of URLs from https://github.com/ada-url/url-various-datasets I ran a script count the number of tags in each document, having previously downloaded all URLs.

Of these, not all downloaded successfully and not all were HTML files.

82,406 HTML files were analyzed, representing pages from the top 100 most popular websites.

When counting all tags, *trunk* took between 310 seconds and 313 seconds across multiple test runs, measured from `microtime()` within the process parsing the HTML, and only measuring around the `next_token()` loop.

On this branch, the counting took between 293 seconds and 300 seconds, representing **around a 5% real-world improvement in token parsing speed**.

For the top100 dataset, this histogram represents the relative parsing speed in MB/s for the branch against `trunk`.

![histogram-of-relative-mbps](https://github.com/WordPress/wordpress-develop/assets/5431237/2f6dbaca-7e08-4ba8-8c28-42d2e57f8d4b)