@dmsnell 2024-07-01
Merged in [[58613]](https://core.trac.wordpress.org/changeset/58613)
https://github.com/wordpress/wordpress-develop/commit/cf064ef249931cfce2af007ec3a95b22b088c191